♻️ members API 분리하고 중복타입 제거 및 API 요청 함수명을 수정했다. - 로그인, 로그아웃, 회원가입 관련 API를 sign 파일에, 멤버조회 관련 API를 members 파일에 분리했다. - MemeberResponseTypes, SignInMemberTypes를 UserTypes로 수정 - generalLogin -> signin, logout -> signout, postSignUpData -> signup으로 함수명 수정
