+++ src/api/members.ts
@@ -1,0 +1,22 @@
+import { UserTypes } from '@/types/issue';
+import axios, { AxiosError } from 'axios';
+
+export const getMemberList = async () => {
+  try {
+    const { data } = await axios.get<UserTypes[]>('api/members');
+    return data;
+  } catch (error) {
+    const err = error as AxiosError;
+    throw err;
+  }
+};
+
+export const getDuplicatesResult = async (router: string, value: string) => {
+  try {
+    const { data } = await axios.get(`api/members/${router}/${value}/exists`);
+    return data;
+  } catch (error) {
+    const err = error as AxiosError;
+    throw err;
+  }
+};
