+++ src/hooks/useFetchSideBarData.ts
@@ -1,11 +1,11 @@
+import { getMemberList } from '@/api/members';
 import { useQuery } from '@tanstack/react-query';
 
-import { getMemberData } from '@/api/login_logout';
 import { getLabelData } from '@/api/labelList';
 import { getMilestoneData } from '@/api/milestone';
 
 const useFetchSideBarData = () => {
-  const { data: memberData } = useQuery(['members'], getMemberData);
+  const { data: memberData } = useQuery(['members'], getMemberList);
   const { data: labelData } = useQuery(['labels'], getLabelData);
   const { data: milestoneData } = useQuery(['milestones'], getMilestoneData);
