+++ src/hooks/useLogout.ts
@@ -1,14 +1,14 @@
 import { useResetRecoilState, useSetRecoilState } from 'recoil';
 import OAuthState from '@/stores/auth';
 import { LoginUserInfoState } from '@/stores/loginUserInfo';
-import { logout } from '@/api/login_logout';
+import { signout } from '@/api/sign';
 
 const useLogout = () => {
   const setIsOAuth = useSetRecoilState(OAuthState);
   const resetLoginUserInfo = useResetRecoilState(LoginUserInfoState);
 
   const setLogout = async () => {
-    await logout();
+    await signout();
     resetLoginUserInfo();
     setIsOAuth(false);
   };
